DadBot 0.3.7
Removed Kick/Ban administration commands. Bot is in too many unique servers for specific roles to be used for administration access. I've kept purge in production for the sole reason that it can be used within my personal server.

Removed Hug, Gottem, Encourage, and a few words from the %speak command.

IP bans for %image have become less frequent though I've changed nothing. Not sure if this is a search engine function or the fact that I moved the bot off AWS and to an on-prem server. I believe it may be the latter. 

Begin planning for the EOL process for this bot. He's under utilized and the upkeep costs (time, stress) are much too high for how little he's getting used.